@@ -46,72 +46,68 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "15.2.3",
-        "@angular/common": "15.2.3",
-        "@angular/compiler": "15.2.3",
-        "@angular/core": "15.2.3",
-        "@angular/forms": "15.2.3",
-        "@angular/platform-browser": "15.2.3",
-        "@angular/platform-browser-dynamic": "15.2.3",
-        "@angular/router": "15.2.3",
-        "@nestjs/common": "9.2.0",
-        "@nestjs/core": "9.2.0",
-        "@nestjs/platform-express": "9.2.0",
-        "@nrwl/angular": "15.8.6",
+        "@angular/animations": "18.2.5",
+        "@angular/common": "18.2.5",
+        "@angular/compiler": "18.2.5",
+        "@angular/core": "18.2.5",
+        "@angular/forms": "18.2.5",
+        "@angular/platform-browser": "18.2.5",
+        "@angular/platform-browser-dynamic": "18.2.5",
+        "@angular/router": "18.2.5",
+        "@nestjs/common": "10.4.3",
+        "@nestjs/core": "10.4.3",
+        "@nestjs/platform-express": "10.4.3",
+        "@nx/angular": "19.7.4",
         "reflect-metadata": "^0.1.13",
         "rxjs": "7.5.7",
         "tslib": "^2.4.1",
-        "zone.js": "0.12.0"
+        "zone.js": "0.14.8"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "15.2.4",
-        "@angular-devkit/core": "15.1.0",
-        "@angular-devkit/schematics": "15.1.0",
-        "@angular-eslint/eslint-plugin": "15.0.0",
-        "@angular-eslint/eslint-plugin-template": "15.0.0",
-        "@angular-eslint/template-parser": "15.0.0",
-        "@angular/cli": "15.2.0",
-        "@angular/compiler-cli": "15.2.3",
-        "@angular/language-service": "15.2.3",
+        "@angular-devkit/build-angular": "18.2.5",
+        "@angular-devkit/core": "18.2.5",
+        "@angular-devkit/schematics": "18.2.5",
+        "@angular-eslint/eslint-plugin": "18.3.1",
+        "@angular-eslint/eslint-plugin-template": "18.3.1",
+        "@angular-eslint/template-parser": "18.3.1",
+        "@angular/cli": "~18.2.0",
+        "@angular/compiler-cli": "18.2.5",
+        "@angular/language-service": "18.2.5",
         "@badeball/cypress-cucumber-preprocessor": "15.1.5",
         "@bahmutov/cypress-esbuild-preprocessor": "2.2.0",
         "@cypress/webpack-dev-server": "2.3.0",
-        "@nestjs/schematics": "9.0.3",
-        "@nestjs/testing": "9.1.4",
-        "@nrwl/cli": "15.2.1",
-        "@nrwl/cypress": "15.8.6",
-        "@nrwl/eslint-plugin-nx": "15.8.6",
-        "@nrwl/jest": "15.8.6",
-        "@nrwl/js": "15.8.6",
-        "@nrwl/linter": "15.8.6",
-        "@nrwl/nest": "15.8.6",
-        "@nrwl/node": "15.8.6",
-        "@nrwl/workspace": "15.8.6",
-        "@schematics/angular": "15.1.0",
-        "@types/jest": "29.4.4",
-        "@types/node": "18.7.21",
+        "@nestjs/schematics": "10.1.4",
+        "@nestjs/testing": "10.4.3",
+        "@nx/cypress": "19.7.4",
+        "@nx/eslint": "19.7.4",
+        "@nx/eslint-plugin": "19.7.4",
+        "@nx/jest": "19.7.4",
+        "@nx/js": "19.7.4",
+        "@nx/nest": "19.7.4",
+        "@nx/node": "19.7.4",
+        "@nx/workspace": "19.7.4",
+        "@schematics/angular": "18.2.5",
+        "@types/jest": "29.5.13",
+        "@types/node": "18.16.9",
         "@types/supertest": "^2.0.12",
-        "@typescript-eslint/eslint-plugin": "5.43.0",
-        "@typescript-eslint/parser": "5.43.0",
-        "cypress": "12.2.0",
+        "@typescript-eslint/eslint-plugin": "7.18.0",
+        "@typescript-eslint/parser": "7.18.0",
+        "@typescript-eslint/utils": "^7.16.0",
+        "cypress": "^13.0.0",
         "dotenv": "16.0.3",
-        "eslint": "8.28.0",
-        "eslint-config-prettier": "8.5.0",
-        "eslint-plugin-cypress": "2.12.1",
+        "eslint": "8.57.1",
+        "eslint-config-prettier": "9.1.0",
+        "eslint-plugin-cypress": "2.15.2",
         "html-webpack-plugin": "5.5.0",
-        "jest": "29.4.3",
-        "jest-environment-jsdom": "29.4.3",
+        "jest": "29.7.0",
+        "jest-environment-jsdom": "29.7.0",
         "jest-extended": "^3.1.0",
-        "jest-preset-angular": "13.0.0",
-        "nx": "15.8.6",
+        "jest-preset-angular": "14.1.1",
+        "nx": "19.7.4",
         "prettier": "2.7.1",
         "supertest": "^6.3.3",
-        "ts-jest": "29.0.5",
+        "ts-jest": "29.1.5",
         "ts-node": "10.9.1",
-        "typescript": "4.9.5"
-    },
-    "engines": {
-        "node": "18.15.0",
-        "npm": "9.5.0"
+        "typescript": "5.5.4"
     }
 }
