@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppComponent } from './app.component';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import {
     FrontendUiComponentsModule,
     SimpleStandaloneComponentComponent,
@@ -12,15 +12,16 @@ import {
     declarations: [
         AppComponent,
     ],
+    bootstrap: [
+        AppComponent,
+    ],
     imports: [
         BrowserModule,
-        HttpClientModule,
         SimpleStandaloneComponentComponent,
         FrontendUiComponentsModule,
     ],
-    providers: [],
-    bootstrap: [
-        AppComponent,
+    providers: [
+        provideHttpClient(withInterceptorsFromDi()),
     ],
 })
 export class AppModule { }
