@@ -1,5 +1,5 @@
 import { defineConfig } from 'cypress';
-import { nxComponentTestingPreset } from '@nrwl/angular/plugins/component-testing';
+import { nxComponentTestingPreset } from '@nx/angular/plugins/component-testing';
 
 export default defineConfig({
     component: nxComponentTestingPreset(__filename),
