@@ -1,7 +1,5 @@
-const { getJestProjects } = require('@nrwl/jest');
+const { getJestProjects } = require('@nx/jest');
 
 export default {
-    projects: [
-        ...getJestProjects(),
-    ],
+    projects: [...getJestProjects()],
 };
